Move buttons and menu to thumbnail footer
